Update SAF_FBK.lp 'Supply Fan Feedback Looping Program
As I pointed out in the Teams chat on 7/29/20, the FailStamp being defined as a simple Numeric will cause an error (at least when the script is run on an i2 or b3 controller in EBO).  The program will immediately go to the E line when attempting to initialize (FailStamp = Date).  This prevents the program from ever getting to FanFbk where SAF_RI is set.  As a result, SAF_RI (or SaFanSts for the new point naming) will never get set and any programs where this signal is required, will not execute as expected.  I simply defined SAF_RI as a Datetime rather than a Numeric and it seems to resolve the issue in EBO.